refactor(dev): route exports util
