+++ packages/remix-dev/compiler/browserjs/plugins/routes_unstable.ts
@@ -8,7 +8,7 @@ import invariant from "../../../invariant";
 import * as Transform from "../../../transform";
 import type { CompileOptions } from "../../options";
 import { getLoaderForFile } from "../../utils/loaders";
-import { getRouteModuleExports } from "../../routeExports";
+import { getRouteModuleExports } from "../../utils/routeExports";
 import { applyHMR } from "./hmr";
 
 const serverOnlyExports = new Set(["action", "loader"]);
