+++ packages/remix-dev/compiler/utils/manifest.ts
@@ -5,7 +5,7 @@ import type * as esbuild from "esbuild";
 import type { RemixConfig } from "../../config";
 import invariant from "../../invariant";
 import { type Manifest } from "../../manifest";
-import { getRouteModuleExports } from "../routeExports";
+import { getRouteModuleExports } from "./routeExports";
 import { getHash } from "./crypto";
 
 type Route = RemixConfig["routes"][string];
