+++ packages/remix-dev/compiler/utils/routeExports.ts
@@ -1,10 +1,10 @@
 import * as path from "path";
 import * as esbuild from "esbuild";
 
-import * as cache from "../cache";
-import type { RemixConfig } from "../config";
-import { mdxPlugin } from "./plugins/mdx";
-import { getFileHash } from "./utils/crypto";
+import * as cache from "../../cache";
+import type { RemixConfig } from "../../config";
+import { mdxPlugin } from "../plugins/mdx";
+import { getFileHash } from "./crypto";
 
 type CachedRouteExports = { hash: string; exports: string[] };
